requirements-docker.txt
# Docker/Railway: torch Dockerfile da CPU wheel bilan alohida o‘rnatiladi
opencv-python-headless>=4.8.0
numpy>=1.24.0
ultralytics>=8.3.0
pillow>=10.0.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pytz>=2023.3
